feat(elections): let party answers carry candidate names
Keeps the party panel — proper party colours, the Other line, the map
colouring — but lets its answers name the nominee, so "Democratic party"
reads "Josh Turek (D)". That removes the duplication where a state showed
the party split and a near-identical "who's running" card underneath.

isDemocraticAnswer / isRepublicanAnswer now also accept a trailing "(D)" /
"(R)" tag. Anchored on the parentheses so "(D)" matches while "(Dem-leaning)"
does not, and "(R)" cannot match inside "(Rep)"; no /g flag, since a sticky
lastIndex would make .test() alternate on repeat calls. Verified additive:
no answer on any market the elections page reads currently contains a
parenthesised one- or two-letter token, so nothing changes until a market is
deliberately renamed.

The head-to-head alignment in party-panel now derives from
isDemocraticAnswer rather than an exact 'Democratic Party' match, so it
survives the rename. The Harris/Trump portraits still use exact matching and
so fall back to a generic icon on renamed answers — which is correct, since
a Harris photo beside "Josh Turek (D)" would be wrong.

Extends the rename script to cover the six competitive Senate party markets
(TX, OH, IA, ME, MI, AK) alongside the House districts, and renames it to
match. Nominees verified 2026-09-07.

⚠️ These markets resolve on PARTY, not on the named candidate — the (D)/(R)
tags are load-bearing, both for meaning and because the matchers read them.
Maine is the live proof that nominees move: Mills quit in April, Platner won
the primary and withdrew in July, and Troy Jackson was nominated on 25 July.
The community candidate market for Maine still lists Mills and Platner and
prices Collins at 99%, while the party market has the race at 67% Democratic
— which is a good argument for the party market being the one on the page.
